refactor: split client construction from inspector construction
Building the client that talks to a provider and building the inspector over
that client are separate concerns, so they get a module each. The crate
resolved credentials and nothing else, so it is renamed for what it builds now.
