+++ crates/foreign-chain-rpc-factory/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "foreign-chain-rpc-auth"
+name = "foreign-chain-rpc-factory"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
