+++ crates/node/Cargo.toml
@@ -25,7 +25,7 @@ ed25519-dalek = { workspace = true }
 flume = { workspace = true }
 foreign-chain-health-check = { workspace = true }
 foreign-chain-inspector = { workspace = true }
-foreign-chain-rpc-auth = { workspace = true }
+foreign-chain-rpc-factory = { workspace = true }
 foreign-chain-rpc-interfaces = { workspace = true }
 futures = { workspace = true }
 gcloud-sdk = { workspace = true }
